Revert "refactor: separate distinct services into new files"
This reverts commit 78af99c2877bd11e5ef225c36b1bf825002970c6.
