+++ src/services/index.js
@@ -1,5 +1,3 @@
-export * from './social-media-service.js'
+export * from './db-service.js'
 export * from './validation-service.js'
 export * from './intersection-service.js'
-export * from './project-service.js'
-export * from './skill-service.js'
