--- src/services/skill-service.js
@@ -1,28 +1,0 @@
-function getPlaceholderSkills() {
-    return [
-        {
-            id: 0,
-            title: "HTML",
-            type: "hard skill",
-            desc: "lorem ipsum dolor sit amet",
-            banner: {
-                path: "https://picsum.photos/200/300",
-                alt: "",
-            },
-        },
-        {
-            id: 1,
-            title: "CSS",
-            type: "hard skill",
-            desc: "lorem ipsum dolor sit amet",
-            banner: {
-                path: "https://picsum.photos/200/300",
-                alt: "",
-            },
-        },
-    ]
-}
-
-export const skillService = {
- getPlaceholderSkills   
-}
