--- src/services/social-media-service.js
@@ -1,16 +1,0 @@
-function getAllSocialLinks() {
-  return [
-    {
-      name: "Github",
-      url: "https://github.com/trsaints",
-    },
-    {
-      name: "LinkedIn",
-      url: "https://www.linkedin.com/in/trsaints2003/",
-    },
-  ]
-}
-
-export const socialMediaService = {
-  getAllSocialLinks,
-}
